Save the Lesson Text and Videos
Instead of just grabbing the Workbook & Phrase/Vocab videos, grab the
text for each Unit and any inline videos as well.

Unit text is saved in markdown as `lesson.md` and the videos are stored
in a `lesson` folder. The phrase and vocab lists are stripped from the
lesson text.

The `lesson` & `phrases` video filenames have been prefixed with a
number indicating the order in which they appear. By sorting by
filename, you can view the videos in the same order as they appear on
the website.

* Refactor phrase & vocab downloading into method.